+++ source/ChanSort.Loader.Sony/SonyPlugin.cs
@@ -1,4 +1,5 @@
-﻿using ChanSort.Api;
+﻿using System.IO;
+using ChanSort.Api;
 
 namespace ChanSort.Loader.Sony
 {
@@ -10,6 +11,14 @@ namespace ChanSort.Loader.Sony
 
     public SerializerBase CreateSerializer(string inputFile)
     {
+      using (var rdr = new StreamReader(inputFile))
+      {
+        var line1 = rdr.ReadLine() ?? "";
+        var line2 = rdr.ReadLine() ?? "";
+        if (line1.Contains("<service_list_transfer>") || line2.Contains("<service_list_transfer>"))
+          return new MediaTek.Serializer(inputFile, true);
+      }
+
       return new Serializer(inputFile);
     }
   }